Implementing the builder pattern
This follow the recipe on Effective Java 2nd.
I did in C++ so I can fixate the construct (private constructor that only has the builder as parameters) by not merely copying the book example.
Not sure if it is the best implementation in the point of view of C++, specially C++11 and C++14 standards. The Builder.build() method forces the copy of the object. I could have return a pointer to instance allocated in the heap ( new NutrictionFacts) , but I wanted to do everything static since Java is always dynamic (heap).